moving code into managable sections
